split to 2 files + baseline for new flow
